% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autolog.R
\docType{package}
\name{autolog-package}
\alias{autolog-package}
\title{Automatic Logging}
\description{
This package provides a one-liner way to add logging to an entire package or R session.
}
\author{
Neal Fultz \email{njf@zestfinance.com}
}
\seealso{
\code{\link{autolog}}, \code{\link{wrap}}
}

