% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/equipTron.R
\name{equipTron}
\alias{equipTron}
\title{Equip a package with tron}
\usage{
equipTron()
}
\description{
Adds a tron startup hook to a packages \code{.onLoad} and \code{DESCRIPTION}.
Run this from a package's source root directory.
}

