% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tron.function.R
\name{.tron.function}
\alias{.troff.function}
\alias{.tron.function}
\title{tron - Wrap a function in logging code}
\usage{
\method{.tron}{function}(x, pre, post = pre)

\method{.troff}{function}(x)
}
\arguments{
\item{x}{a function to decorate}

\item{pre}{a function, to be called before \code{f}}

\item{post}{a function, to be called after \code{f}}
}
\description{
Create a logged copy of a function. Every time the new function is called, all three functions are called in order:
\enumerate{ 
\item \code{pre}
\item \code{f}
\item \code{post}
}
}
\details{
Wrapped functions carry a \dQuote{tron} attribute, which can be tested for
using \code{is.tron}. The original function \code{f} can be extracted
using \code{troff}.
}
\examples{
f <- tron(sum, message)
f(1:10)
is.tron(f)
f <- troff(f)
f(1:10)
}
\seealso{
\url{http://en.wikipedia.org/wiki/Decorator_pattern} and  \code{\link[memoise]{memoise}} for another example of \dQuote{decorator} functions.
}

