% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autolog.R
\name{wrap}
\alias{is.autologged}
\alias{unwrap}
\alias{wrap}
\title{Wrap a function in logging code}
\usage{
wrap(f, pre, post = pre)

is.autologged(f)

unwrap(f)
}
\arguments{
\item{f}{a function to decorate}

\item{pre}{a function, to be called before \code{f}}

\item{post}{a function, to be called after \code{f}}
}
\description{
Create a logged copy of a function. Every time the new function is called, all three functions are called in order:
\enumerate{
\item \code{pre}
\item \code{f}
\item \code{post}
}
}
\details{
Wrapped functions carry an \dQuote{autologged} attribute, which can be tested for using \code{is.autologged}. The original function \code{f} can be extracted
using \code{unwrap}.
}
\examples{
f <- wrap(sum, message)
f(1:10)
is.autologged(f)
f <- unwrap(f)
f(1:10)
}
\seealso{
\url{http://en.wikipedia.org/wiki/Decorator_pattern} and  \code{\link[memoise]{memoise}} for another example of \dQuote{decorator} functions.
}

