% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autolog.R
\name{autolog}
\alias{autolog}
\title{Automatic Logging}
\usage{
autolog(e = .GlobalEnv, logger = "message",
  verbose = getOption("autolog.verbose", FALSE))
}
\arguments{
\item{e}{an environment to process; default to Global}

\item{logger}{a logging function accepting \code{...}}

\item{verbose}{verbose output of what functions are being instrumented}
}
\description{
Automatic Logging
}
\examples{
f <- function(a,b) a / b
zzz <- function(x,y) f(x,y) / f(y,x)
autolog(environment(), verbose=TRUE)
zzz(2,1)
}

