% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tron.environment.R
\name{.tron.environment}
\alias{.troff.environment}
\alias{.tron.environment}
\title{tron - Log a session or package}
\usage{
\method{.tron}{environment}(x, logger = getOption("tron.logger", "message"),
  verbose = getOption("tron.verbose", FALSE))

\method{.troff}{environment}(x)
}
\arguments{
\item{x}{an environment to process; defaults to the \code{\link{.GlobalEnv}}}

\item{logger}{a logging function or name of function which accepts \code{...}}

\item{verbose}{logical, log which functions are detected and modified}
}
\description{
Call \code{\link{.tron.function}} on each function in an environment and assign the result back.
}
\section{Options}{

You can set the following default parameters using \code{\link{option}}:
\describe{
\item{tron.logger}{ name of a logging function}
\item{tron.verbose}{ logical }
}
}
\examples{
f <- function(a,b) a / b
zzz <- function(x,y) f(x,y) / f(y,x)
tron(environment(), verbose=TRUE)
zzz(2,1)
}

