% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tron.R
\docType{package}
\name{tron}
\alias{is.tron}
\alias{tron}
\alias{tron-package}
\alias{untron}
\title{tron - Automatic Logging}
\usage{
tron(x, ...)

untron(x, ...)

is.tron(f)
}
\description{
This package provides a one-liner way to add logging to a function,
session, or an entire package.
}
\author{
Neal Fultz \email{njf@zestfinance.com}
}
\seealso{
\code{\link{tron.function}}, \code{\link{tron.environment}}
}

