% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tron.R
\docType{package}
\name{tron}
\alias{is.tron}
\alias{troff}
\alias{tron}
\alias{tron-package}
\title{tron - Automatic Logging}
\usage{
tron(x, ...)

troff(x, ...)

is.tron(x)
}
\arguments{
\item{x}{an object}

\item{...}{Arguments to be passed to methods.}
}
\description{
This package provides a one-liner way to enable/disable logging
to a function, session, or an entire package.
}
\author{
Neal Fultz \email{njf@zestfinance.com}
}
\seealso{
\code{\link{tron.function}}, \code{\link{tron.environment}}
}

